+++ packages/core/src/components/rtk-meeting/rtk-meeting.tsx
@@ -18,6 +18,7 @@ import {
   type RtkUiStoreExtended,
 } from '../../utils/sync-with-store/ui-store';
 import { Overrides, defaultOverrides } from '../../lib/overrides';
+import { getInitErrorInfo } from '../../utils/init-error';
 import { getJoinErrorInfo } from '../../utils/join-error';
 
 export type MeetingMode = 'fixed' | 'fill';
@@ -43,7 +44,7 @@ export class RtkMeeting {
   private providerId: string = 'provider-' + Math.floor(Math.random() * 100);
 
   private roomJoinedListener = () => {
-    this.updateStates({ meeting: 'joined', joinError: undefined, joinErrorCode: undefined });
+    this.updateStates({ meeting: 'joined', preJoinError: null });
   };
 
   private waitlistedListener = () => {
@@ -131,16 +132,15 @@ export class RtkMeeting {
    */
   @Event({ eventName: 'rtkStatesUpdate' }) statesUpdate: EventEmitter<States>;
 
-  private authErrorListener: (ev: CustomEvent<Error>) => void;
+  private initErrorListener: (ev: CustomEvent) => void;
 
   connectedCallback() {
     if (typeof window !== 'undefined') {
-      this.authErrorListener = (ev) => {
-        if (ev.detail.message.includes('401')) {
-          this.updateStates({ meeting: 'ended', roomLeftState: 'unauthorized' });
-        }
+      this.initErrorListener = (ev) => {
+        const { message, code } = getInitErrorInfo(this.t, ev.detail);
+        this.updateStates({ preJoinError: { message, code } });
       };
-      window.addEventListener('rtkError', this.authErrorListener);
+      window.addEventListener('ClientError', this.initErrorListener);
     }
 
     // Initialize default values
@@ -175,7 +175,7 @@ export class RtkMeeting {
       this.meeting?.leave();
     }
     this.resizeObserver.disconnect();
-    window.removeEventListener('rtkError', this.authErrorListener);
+    window.removeEventListener('ClientError', this.initErrorListener);
 
     // Remove event listeners
     if (this.storeRequestListener) {
@@ -323,16 +323,16 @@ export class RtkMeeting {
         meeting
           .join()
           .then(() => {
-            this.updateStates({ joinError: undefined, joinErrorCode: undefined });
+            this.updateStates({ preJoinError: null });
           })
           .catch((err) => {
             const { message, code } = getJoinErrorInfo(this.t, err);
-            this.updateStates({ joinError: message, joinErrorCode: code });
+            this.updateStates({ preJoinError: { message, code } });
           });
       }
     }
 
-    window.removeEventListener('rtkError', this.authErrorListener);
+    window.removeEventListener('ClientError', this.initErrorListener);
   }
 
   private handleChangingMeeting = (destinationMeetingId: string) => {